+++ test/functional-portable/cli/noncloud/cli_test.go
@@ -44,8 +44,9 @@ import (
 	"github.com/radius-project/radius/pkg/corerp/api/v20231001preview"
 	"github.com/radius-project/radius/pkg/ucp/resources"
 	"github.com/radius-project/radius/pkg/version"
-	"github.com/radius-project/radius/test/functional-portable/corerp"
+
 	"github.com/radius-project/radius/test/radcli"
+	"github.com/radius-project/radius/test/rp"
 	"github.com/radius-project/radius/test/step"
 	"github.com/radius-project/radius/test/testcontext"
 	"github.com/radius-project/radius/test/testutil"
@@ -59,8 +60,8 @@ const (
 	retries = 10
 )
 
-func verifyRecipeCLI(ctx context.Context, t *testing.T, test corerp.RPTest) {
-	options := corerp.NewRPTestOptions(t)
+func verifyRecipeCLI(ctx context.Context, t *testing.T, test rp.RPTest) {
+	options := rp.NewRPTestOptions(t)
 	cli := radcli.NewCLI(t, options.ConfigFilePath)
 	envName := test.Steps[0].RPResources.Resources[0].Name
 	registry := strings.TrimPrefix(testutil.GetBicepRecipeRegistry(), "registry=")
@@ -155,8 +156,8 @@ func verifyRecipeCLI(ctx context.Context, t *testing.T, test corerp.RPTest) {
 	})
 }
 
-func verifyCLIBasics(ctx context.Context, t *testing.T, test corerp.RPTest) {
-	options := corerp.NewRPTestOptions(t)
+func verifyCLIBasics(ctx context.Context, t *testing.T, test rp.RPTest) {
+	options := rp.NewRPTestOptions(t)
 	cli := radcli.NewCLI(t, options.ConfigFilePath)
 	appName := test.Name
 	containerName := "containerA"
@@ -280,7 +281,7 @@ func Test_Run_Logger(t *testing.T) {
 	// Will be used to cancel `rad run`
 	ctx, cancel := testcontext.NewWithCancel(t)
 	t.Cleanup(cancel)
-	options := corerp.NewRPTestOptions(t)
+	options := rp.NewRPTestOptions(t)
 
 	template := "testdata/corerp-kubernetes-cli-run.bicep"
 	applicationName := "cli-run-logger"
@@ -354,7 +355,7 @@ func Test_Run_Portforward(t *testing.T) {
 	// Will be used to cancel `rad run`
 	ctx, cancel := testcontext.NewWithCancel(t)
 	t.Cleanup(cancel)
-	options := corerp.NewRPTestOptions(t)
+	options := rp.NewRPTestOptions(t)
 
 	template := "testdata/corerp-kubernetes-cli-run-portforward.bicep"
 	applicationName := "cli-run-portforward"
@@ -453,7 +454,7 @@ func Test_CLI(t *testing.T) {
 	template := "testdata/corerp-kubernetes-cli.bicep"
 	name := "kubernetes-cli"
 
-	test := corerp.NewRPTest(t, name, []corerp.TestStep{
+	test := rp.NewRPTest(t, name, []rp.TestStep{
 		{
 			Executor: step.NewDeployExecutor(template, testutil.GetMagpieImage()),
 			RPResources: &validation.RPResourceSet{
@@ -493,7 +494,7 @@ func Test_CLI_JSON(t *testing.T) {
 	template := "testdata/corerp-kubernetes-cli.json"
 	name := "kubernetes-cli-json"
 
-	test := corerp.NewRPTest(t, name, []corerp.TestStep{
+	test := rp.NewRPTest(t, name, []rp.TestStep{
 		{
 			Executor: step.NewDeployExecutor(template, testutil.GetMagpieImage()),
 			RPResources: &validation.RPResourceSet{
@@ -533,7 +534,7 @@ func Test_CLI_Delete(t *testing.T) {
 	ctx, cancel := testcontext.NewWithCancel(t)
 	t.Cleanup(cancel)
 
-	options := corerp.NewRPTestOptions(t)
+	options := rp.NewRPTestOptions(t)
 	appName := "kubernetes-cli-with-resources"
 	appNameUnassociatedResources := "kubernetes-cli-with-unassociated-resources"
 	appNameEmptyResources := "kubernetes-cli-empty-resources"
@@ -631,7 +632,7 @@ func Test_CLI_DeploymentParameters(t *testing.T) {
 	// corerp-kubernetes-cli-parameters.parameters.json uses ghcr.io/radius-project/dev as a registry parameter.
 	// Use the specified tag only if the test uses ghcr.io/radius-project/dev registry. Otherwise, use latest tag.
 
-	test := corerp.NewRPTest(t, name, []corerp.TestStep{
+	test := rp.NewRPTest(t, name, []rp.TestStep{
 		{
 			Executor: step.NewDeployExecutor(template, "@"+parameterFilePath, testutil.GetMagpieTag()),
 			RPResources: &validation.RPResourceSet{
@@ -670,7 +671,7 @@ func Test_CLI_version(t *testing.T) {
 	ctx, cancel := testcontext.NewWithCancel(t)
 	t.Cleanup(cancel)
 
-	options := corerp.NewTestOptions(t)
+	options := rp.NewTestOptions(t)
 	cli := radcli.NewCLI(t, options.ConfigFilePath)
 
 	output, err := cli.Version(ctx)
@@ -691,7 +692,7 @@ func Test_CLI_Only_version(t *testing.T) {
 	ctx, cancel := testcontext.NewWithCancel(t)
 	t.Cleanup(cancel)
 
-	options := corerp.NewTestOptions(t)
+	options := rp.NewTestOptions(t)
 	cli := radcli.NewCLI(t, options.ConfigFilePath)
 
 	output, err := cli.CliVersion(ctx)
@@ -709,7 +710,7 @@ func Test_RecipeCommands(t *testing.T) {
 	template := "testdata/corerp-resources-recipe-env.bicep"
 	name := "corerp-resources-recipe-env"
 
-	test := corerp.NewRPTest(t, name, []corerp.TestStep{
+	test := rp.NewRPTest(t, name, []rp.TestStep{
 		{
 			Executor: step.NewDeployExecutor(template, testutil.GetBicepRecipeRegistry(), testutil.GetBicepRecipeVersion()),
 			RPResources: &validation.RPResourceSet{
@@ -735,7 +736,7 @@ func Test_DevRecipes(t *testing.T) {
 	ctx, cancel := testcontext.NewWithCancel(t)
 	t.Cleanup(cancel)
 
-	options := corerp.NewTestOptions(t)
+	options := rp.NewTestOptions(t)
 	cli := radcli.NewCLI(t, options.ConfigFilePath)
 
 	envName := "test-dev-recipes"
@@ -797,7 +798,7 @@ func GetAvailablePort() (int, error) {
 
 // DeleteAppWithoutDeletingResources creates a client to delete an application without deleting its resources and returns
 // an error if one occurs.
-func DeleteAppWithoutDeletingResources(t *testing.T, ctx context.Context, options corerp.RPTestOptions, applicationName string) error {
+func DeleteAppWithoutDeletingResources(t *testing.T, ctx context.Context, options rp.RPTestOptions, applicationName string) error {
 	client := options.ManagementClient
 	require.IsType(t, client, &clients.UCPApplicationsManagementClient{})
 	appManagementClient := client.(*clients.UCPApplicationsManagementClient)
