+++ test/functional-portable/corerp/cloud/resources/aws_multi_identifier_resource_test.go
@@ -22,7 +22,7 @@ import (
 	"time"
 
 	"github.com/google/uuid"
-	"github.com/radius-project/radius/test/functional-portable/corerp"
+	"github.com/radius-project/radius/test/rp"
 	"github.com/radius-project/radius/test/step"
 	"github.com/radius-project/radius/test/validation"
 )
@@ -34,7 +34,7 @@ func Test_AWS_MultiIdentifier_Resource(t *testing.T) {
 	testName := "ms" + uuid.New().String()
 	creationTimestamp := fmt.Sprintf("%d", time.Now().Unix())
 
-	test := corerp.NewRPTest(t, testName, []corerp.TestStep{
+	test := rp.NewRPTest(t, testName, []rp.TestStep{
 		{
 			Executor:                               step.NewDeployExecutor(template, "filterName="+filterName, "logGroupName="+logGroupName, "creationTimestamp="+creationTimestamp),
 			SkipKubernetesOutputResourceValidation: true,
@@ -64,6 +64,6 @@ func Test_AWS_MultiIdentifier_Resource(t *testing.T) {
 		},
 	})
 
-	test.RequiredFeatures = []corerp.RequiredFeature{corerp.FeatureAWS}
+	test.RequiredFeatures = []rp.RequiredFeature{rp.FeatureAWS}
 	test.Test(t)
 }
