+++ test/functional-portable/corerp/noncloud/resources/container_test.go
@@ -21,7 +21,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/radius-project/radius/test/functional-portable/corerp"
+	"github.com/radius-project/radius/test/rp"
 	"github.com/radius-project/radius/test/step"
 	"github.com/radius-project/radius/test/testutil"
 	"github.com/radius-project/radius/test/validation"
@@ -34,7 +34,7 @@ func Test_Container(t *testing.T) {
 	name := "corerp-resources-container"
 	appNamespace := "corerp-resources-container-app"
 
-	test := corerp.NewRPTest(t, name, []corerp.TestStep{
+	test := rp.NewRPTest(t, name, []rp.TestStep{
 		{
 			Executor: step.NewDeployExecutor(template, testutil.GetMagpieImage()),
 			RPResources: &validation.RPResourceSet{
@@ -68,7 +68,7 @@ func Test_ContainerDNSSD_TwoContainersDNS(t *testing.T) {
 	name := "corerp-resources-container-two-containers-dns"
 	appNamespace := "corerp-resources-container-two-containers-dns"
 
-	test := corerp.NewRPTest(t, name, []corerp.TestStep{
+	test := rp.NewRPTest(t, name, []rp.TestStep{
 		{
 			Executor: step.NewDeployExecutor(template, testutil.GetMagpieImage()),
 			RPResources: &validation.RPResourceSet{
@@ -109,7 +109,7 @@ func Test_ContainerDNSSD_OptionalPortScheme(t *testing.T) {
 	name := "corerp-resources-container-optional-port-scheme"
 	appNamespace := "corerp-resources-container-optional-port-scheme"
 
-	test := corerp.NewRPTest(t, name, []corerp.TestStep{
+	test := rp.NewRPTest(t, name, []rp.TestStep{
 		{
 			Executor: step.NewDeployExecutor(template, testutil.GetMagpieImage()),
 			RPResources: &validation.RPResourceSet{
@@ -158,7 +158,7 @@ func Test_ContainerReadinessLiveness(t *testing.T) {
 	name := "corerp-resources-container-live-ready"
 	appNamespace := "corerp-resources-container-live-ready-app"
 
-	test := corerp.NewRPTest(t, name, []corerp.TestStep{
+	test := rp.NewRPTest(t, name, []rp.TestStep{
 		{
 			Executor: step.NewDeployExecutor(template, testutil.GetMagpieImage()),
 			RPResources: &validation.RPResourceSet{
@@ -192,7 +192,7 @@ func Test_ContainerManualScale(t *testing.T) {
 	name := "corerp-resources-container-manualscale"
 	appNamespace := "corerp-resources-container-manualscale-app"
 
-	test := corerp.NewRPTest(t, name, []corerp.TestStep{
+	test := rp.NewRPTest(t, name, []rp.TestStep{
 		{
 			Executor: step.NewDeployExecutor(template, testutil.GetMagpieImage()),
 			RPResources: &validation.RPResourceSet{
@@ -226,7 +226,7 @@ func Test_ContainerWithCommandAndArgs(t *testing.T) {
 	name := "corerp-resources-container-cmd-args"
 	appNamespace := "corerp-resources-container-cmd-args-app"
 
-	test := corerp.NewRPTest(t, name, []corerp.TestStep{
+	test := rp.NewRPTest(t, name, []rp.TestStep{
 		{
 			Executor: step.NewDeployExecutor(container),
 			RPResources: &validation.RPResourceSet{
@@ -249,7 +249,7 @@ func Test_ContainerWithCommandAndArgs(t *testing.T) {
 					},
 				},
 			},
-			PostStepVerify: func(ctx context.Context, t *testing.T, test corerp.RPTest) {
+			PostStepVerify: func(ctx context.Context, t *testing.T, test rp.RPTest) {
 				label := fmt.Sprintf("radapp.io/application=%s", name)
 				pods, err := test.Options.K8sClient.CoreV1().Pods(appNamespace).List(ctx, metav1.ListOptions{
 					LabelSelector: label,
@@ -299,7 +299,7 @@ func Test_Container_FailDueToNonExistentImage(t *testing.T) {
 		},
 	})
 
-	test := corerp.NewRPTest(t, name, []corerp.TestStep{
+	test := rp.NewRPTest(t, name, []rp.TestStep{
 		{
 			Executor:                               step.NewDeployErrorExecutor(template, validate, "magpieimage=non-existent-image"),
 			SkipKubernetesOutputResourceValidation: true,
@@ -334,7 +334,7 @@ func Test_Container_FailDueToBadHealthProbe(t *testing.T) {
 		},
 	})
 
-	test := corerp.NewRPTest(t, name, []corerp.TestStep{
+	test := rp.NewRPTest(t, name, []rp.TestStep{
 		{
 			Executor:                               step.NewDeployErrorExecutor(template, validate, testutil.GetMagpieImage()),
 			SkipKubernetesOutputResourceValidation: true,
