+++ test/functional-portable/corerp/noncloud/resources/container_versioning_test.go
@@ -21,7 +21,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/radius-project/radius/test/functional-portable/corerp"
+	"github.com/radius-project/radius/test/rp"
 	"github.com/radius-project/radius/test/step"
 	"github.com/radius-project/radius/test/testutil"
 	"github.com/radius-project/radius/test/validation"
@@ -37,7 +37,7 @@ func Test_ContainerVersioning(t *testing.T) {
 	name := "corerp-resources-container-versioning"
 	appNamespace := "default-corerp-resources-container-versioning"
 
-	test := corerp.NewRPTest(t, name, []corerp.TestStep{
+	test := rp.NewRPTest(t, name, []rp.TestStep{
 		{
 			Executor: step.NewDeployExecutor(containerV1, testutil.GetMagpieImage()),
 			RPResources: &validation.RPResourceSet{
@@ -61,7 +61,7 @@ func Test_ContainerVersioning(t *testing.T) {
 				},
 			},
 			SkipResourceDeletion: true,
-			PostStepVerify: func(ctx context.Context, t *testing.T, test corerp.RPTest) {
+			PostStepVerify: func(ctx context.Context, t *testing.T, test rp.RPTest) {
 				label := fmt.Sprintf("radapp.io/application=%s", name)
 				secrets, err := test.Options.K8sClient.CoreV1().Secrets(appNamespace).List(ctx, metav1.ListOptions{
 					LabelSelector: label,
@@ -92,7 +92,7 @@ func Test_ContainerVersioning(t *testing.T) {
 					},
 				},
 			},
-			PostStepVerify: func(ctx context.Context, t *testing.T, test corerp.RPTest) {
+			PostStepVerify: func(ctx context.Context, t *testing.T, test rp.RPTest) {
 				label := fmt.Sprintf("radapp.io/application=%s", name)
 				secrets, err := test.Options.K8sClient.CoreV1().Secrets(appNamespace).List(ctx, metav1.ListOptions{
 					LabelSelector: label,
