+++ test/functional-portable/corerp/noncloud/resources/simulated_environment_test.go
@@ -21,7 +21,7 @@ import (
 	"fmt"
 	"testing"
 
-	"github.com/radius-project/radius/test/functional-portable/corerp"
+	"github.com/radius-project/radius/test/rp"
 	"github.com/radius-project/radius/test/step"
 	"github.com/radius-project/radius/test/testutil"
 	"github.com/radius-project/radius/test/validation"
@@ -34,7 +34,7 @@ func Test_Deployment_SimulatedEnv(t *testing.T) {
 	name := "corerp-resources-simulatedenv"
 	appNamespace := "default-corerp-resources-simulatedenv"
 
-	test := corerp.NewRPTest(t, name, []corerp.TestStep{
+	test := rp.NewRPTest(t, name, []rp.TestStep{
 		{
 			Executor: step.NewDeployExecutor(template, testutil.GetMagpieImage()),
 			RPResources: &validation.RPResourceSet{
@@ -72,7 +72,7 @@ func Test_Deployment_SimulatedEnv(t *testing.T) {
 			},
 			SkipKubernetesOutputResourceValidation: true,
 			SkipObjectValidation:                   true,
-			PostStepVerify: func(ctx context.Context, t *testing.T, ct corerp.RPTest) {
+			PostStepVerify: func(ctx context.Context, t *testing.T, ct rp.RPTest) {
 				// Get pods in app namespace
 				label := fmt.Sprintf("radapp.io/application=%s", name)
 				pods, err := ct.Options.K8sClient.CoreV1().Pods(appNamespace).List(ctx, metav1.ListOptions{
