+++ test/functional-portable/daprrp/noncloud/resources/dapr_serviceinvocation_test.go
@@ -19,7 +19,7 @@ package resource_test
 import (
 	"testing"
 
-	"github.com/radius-project/radius/test/functional-portable/corerp"
+	"github.com/radius-project/radius/test/rp"
 	"github.com/radius-project/radius/test/step"
 	"github.com/radius-project/radius/test/testutil"
 	"github.com/radius-project/radius/test/validation"
@@ -30,7 +30,7 @@ func Test_DaprServiceInvocation(t *testing.T) {
 	name := "dapr-serviceinvocation"
 	appNamespace := "default-dapr-serviceinvocation"
 
-	test := corerp.NewRPTest(t, name, []corerp.TestStep{
+	test := rp.NewRPTest(t, name, []rp.TestStep{
 		{
 			Executor: step.NewDeployExecutor(template, testutil.GetMagpieImage()),
 			RPResources: &validation.RPResourceSet{
@@ -61,7 +61,7 @@ func Test_DaprServiceInvocation(t *testing.T) {
 			},
 		},
 	})
-	test.RequiredFeatures = []corerp.RequiredFeature{corerp.FeatureDapr}
+	test.RequiredFeatures = []rp.RequiredFeature{rp.FeatureDapr}
 
 	test.Test(t)
 }
